fix: almost working
There is one case where the player is facing a direction and you walk in the other, then it will help you walk along the block to get around even if it requires more than half a block walking help. So it is using the direction of the player to increase the 'corner' cutting helper mechanic. Other than this the feel seems to be good now.