fix: make checking target positions for throws/jumps be more robust
Having checks that you are close enough to target can lead to strange rounding bugs or janky animation if the check is too generous. Instead check that we are falling and if not moving past target (distance increasing) then assume we are at target. This ensures that you can never miss and fall out of the map.