perf: only check collisions and movement detection close to relevant entity
This is a first rought implementation of efficient collision detection. For now the grid entities are split into and looked up from is the same as the block size. Movement is done by having a function for getting any entity close to the player/bomb. The groups concept is from pygame. In the future we should try to expand it to handle cleaning up entities that are no longer used. Right now this is not unified and the collision to cleanup things from 1 ton blocks is not nice